Move `Part` to new file, add Types folder
